炸弹（1.2.0）优化了栈的消耗，更换了炸弹的爆炸范围

+++ rt1064/project/code/assigned_box_planner_greedy.h
@@ -9,8 +9,8 @@ extern "C" {
 #endif
 
 typedef struct {
-  int row;
-  int col;
+  int8_t row;
+  int8_t col;
 } Point;
 
 typedef struct {
